# Conflicts:
#	Anytype.xcodeproj/project.pbxproj
#	Anytype/Resources/Strings/en.lproj/Localizable.strings
#	Anytype/Sources/PresentationLayer/Assemblies/ModulesDI.swift
#	Anytype/Sources/PresentationLayer/Assemblies/ModulesDIProtocol.swift
#	Anytype/Sources/PresentationLayer/ShareView/ShareViewModelInteractor.swift
#	Anytype/Sources/ServiceLayer/Block/BlockActionHandler/BlockActionService/BlockActionService.swift
#	Anytype/Sources/ServiceLayer/ServiceLocator.swift
#	Modules/Services/Sources/Models/Details/FileDetails.swift
#	Modules/Services/Sources/Models/Extensions/DetailsLayout+Visible.swift
#	Modules/Services/Sources/Services/BlockActionsServiceSingle/BlockActionsServiceSingle.swift
#	Modules/Services/Sources/Services/BlockActionsServiceSingle/BlockActionsServiceSingleProtocol.swift

+++ Scripts/middle-install.sh
@@ -24,10 +24,23 @@ else
 fi
 
 curl https://maven.pkg.github.com/anyproto/anytype-heart/io.anyproto/anytype-heart-ios/${MIDDLE_VERSION}/anytype-heart-ios-${MIDDLE_VERSION}.gz -f --header "Authorization: token ${token}" -L --output ${PROJECT_DIR}/lib.gz
-rm -rf ${PROJECT_DIR}/Dependencies/Middleware/*
-mkdir -p ${PROJECT_DIR}/Dependencies/Middleware
-tar xzf lib.gz -C ${PROJECT_DIR}/Dependencies/Middleware
-rm lib.gz
 
-rm -rf ${PROJECT_DIR}/Modules/ProtobufMessages/Sources/Protocol/*
-cp -r ${PROJECT_DIR}/Dependencies/Middleware/protobuf/*.swift ${PROJECT_DIR}/Modules/ProtobufMessages/Sources/Protocol
+if [ $?  -eq  0 ]
+then
+    rm -rf ${PROJECT_DIR}/Dependencies/Middleware/*
+    mkdir -p ${PROJECT_DIR}/Dependencies/Middleware
+    tar xzf lib.gz -C ${PROJECT_DIR}/Dependencies/Middleware
+    rm lib.gz
+
+    rm -rf ${PROJECT_DIR}/Modules/ProtobufMessages/Sources/Protocol/*
+    cp -r ${PROJECT_DIR}/Dependencies/Middleware/protobuf/*.swift ${PROJECT_DIR}/Modules/ProtobufMessages/Sources/Protocol
+
+    GREEN='\033[0;32m'
+    printf "${GREEN}Success"
+    echo -e "\a" # play sound
+else
+    RED='\033[0;31m'
+    TERMINATOR='\n\e[0m'
+    printf "${RED}Error downloading middleware, check out token provided${TERMINATOR}"
+    printf "use \"make change-github-token\" command to update token"
+fi
